r/kubernetes_cluster: moving the rest of the functions into parse/validate since these now exist
