+++ azurerm/internal/services/containers/kubernetes_cluster_validate.go
