+++ azurerm/internal/services/containers/parse/kubernetes_node_pool_id.go
