+++ azurerm/internal/services/containers/tests/resource_arm_kubernetes_cluster_node_pool_test.go
@@ -13,7 +13,7 @@ import (
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/acceptance"
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/clients"
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/features"
-	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/services/containers"
+	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/services/containers/parse"
 )
 
 func TestAccAzureRMKubernetesClusterNodePool_autoScale(t *testing.T) {
@@ -602,7 +602,7 @@ func testCheckAzureRMKubernetesClusterNodePoolDestroy(s *terraform.State) error 
 
 		name := rs.Primary.Attributes["name"]
 		kubernetesClusterId := rs.Primary.Attributes["kubernetes_cluster_id"]
-		parsedK8sId, err := containers.ParseKubernetesClusterID(kubernetesClusterId)
+		parsedK8sId, err := parse.KubernetesClusterID(kubernetesClusterId)
 		if err != nil {
 			return fmt.Errorf("Error parsing kubernetes cluster id: %+v", err)
 		}
@@ -634,7 +634,7 @@ func testCheckAzureRMKubernetesNodePoolExists(resourceName string) resource.Test
 
 		name := rs.Primary.Attributes["name"]
 		kubernetesClusterId := rs.Primary.Attributes["kubernetes_cluster_id"]
-		parsedK8sId, err := containers.ParseKubernetesClusterID(kubernetesClusterId)
+		parsedK8sId, err := parse.KubernetesClusterID(kubernetesClusterId)
 		if err != nil {
 			return fmt.Errorf("Error parsing kubernetes cluster id: %+v", err)
 		}
@@ -665,7 +665,7 @@ func testCheckAzureRMKubernetesNodePoolNodeLabels(resourceName string, expectedL
 
 		name := rs.Primary.Attributes["name"]
 		kubernetesClusterId := rs.Primary.Attributes["kubernetes_cluster_id"]
-		parsedK8sId, err := containers.ParseKubernetesClusterID(kubernetesClusterId)
+		parsedK8sId, err := parse.KubernetesClusterID(kubernetesClusterId)
 		if err != nil {
 			return fmt.Errorf("Error parsing kubernetes cluster id: %+v", err)
 		}
